refactor: Create package and move
