+++ src/test/java/lotto/domain/BuyerTest.java
@@ -1,5 +1,9 @@
-package lotto;
+package lotto.domain;
 
+import lotto.domain.Buyer;
+import lotto.domain.BuyerResult;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoTicket;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
